# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "EngineMetadata": {
        "additionalProperties": true,
        "description": "Metadata about the model engine, data mode, and claim limitations.",
        "properties": {
          "claim_limitations": {
            "default": "Binary benign-versus-malicious risk only (learned); stage estimate is rule-derived evidence score; target ranking is NetworkX graph-ranked; isolation is simulated estimate - not causal proof",
            "description": "Claim limitations - distinguishes learned, rule-derived, graph-ranked, simulated, and measured",
            "examples": [
              "Binary benign-versus-malicious risk only (learned); stage estimate is rule-derived evidence score; target ranking is NetworkX graph-ranked; isolation is simulated estimate - not causal proof"
            ],
            "title": "Claim Limitations",
            "type": "string"
          },
          "data_mode": {
            "default": "synthetic",
            "description": "Data mode - synthetic or downloaded",
            "examples": [
              "synthetic"
            ],
            "title": "Data Mode",
            "type": "string"
          },
          "data_source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "https://www.unb.ca/cic/datasets/ids-2017.html",
            "description": "Dataset source URL",
            "examples": [
              "https://www.unb.ca/cic/datasets/ids-2017.html"
            ],
            "title": "Data Source Url"
          },
          "feature_count": {
            "default": 80,
            "description": "Number of model features (ordered)",
            "examples": [
              80
            ],
            "title": "Feature Count",
            "type": "integer"
          },
          "feature_schema_version": {
            "default": "phase03-v1-sealed",
            "description": "Feature schema version",
            "examples": [
              "phase03-v1-sealed"
            ],
            "title": "Feature Schema Version",
            "type": "string"
          },
          "mitre_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "13.1",
            "description": "Pinned MITRE ATT&CK version",
            "examples": [
              "13.1"
            ],
            "title": "Mitre Version"
          },
          "model_family": {
            "default": "random_forest",
            "description": "Model family - learned component (random_forest or logistic_regression)",
            "examples": [
              "random_forest"
            ],
            "title": "Model Family",
            "type": "string"
          },
          "model_version": {
            "default": "phase04-v1",
            "description": "Model version identifier",
            "examples": [
              "phase04-v1"
            ],
            "title": "Model Version",
            "type": "string"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 0.45,
            "description": "Warning threshold chosen on Thursday validation (frozen)",
            "examples": [
              0.45
            ],
            "title": "Threshold"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0.1.0",
            "description": "Application version",
            "examples": [
              "0.1.0"
            ],
            "title": "Version"
          }
        },
        "title": "EngineMetadata",
        "type": "object"
      },
...
      "FlowSummary": {
        "additionalProperties": true,
        "properties": {
          "avg_bytes_per_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Mean Flow Bytes/s observed",
            "examples": [
              5120.0
            ],
            "title": "Avg Bytes Per Sec"
          },
          "avg_packets_per_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Mean Flow Packets/s observed",
            "examples": [
              48.0
            ],
            "title": "Avg Packets Per Sec"
          },
          "distinct_dest_ips": {
            "description": "Distinct destination hosts observed",
            "examples": [
              42
            ],
            "title": "Distinct Dest Ips",
            "type": "integer"
          },
          "distinct_ports": {
            "description": "Distinct destination ports observed",
            "examples": [
              18
            ],
            "title": "Distinct Ports",
            "type": "integer"
          },
          "flow_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias for total_flows",
            "examples": [
              190
            ],
            "title": "Flow Count"
          },
          "predicted_malicious_count": {
            "description": "Predicted malicious count model-predicted (threshold 0.45, 80 features) - not label-derived",
            "examples": [
              12
            ],
            "title": "Predicted Malicious Count",
            "type": "integer"
          },
          "predicted_malicious_ratio": {
            "description": "Predicted malicious ratio model-predicted - not label-derived",
            "examples": [
              0.06
            ],
            "title": "Predicted Malicious Ratio",
            "type": "number"
          },
          "predicted_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source label for predicted fields",
            "examples": [
              "model-predicted (Random Forest, threshold 0.45, 80 features, train-only)"
            ],
            "title": "Predicted Source"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Threshold used",
            "examples": [
              0.45
            ],
            "title": "Threshold"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Timestamp observed",
            "examples": [
              "2017-07-03 09:15:23"
            ],
            "title": "Timestamp"
          },
          "timestamp_iso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO timestamp",
            "examples": [
              "2017-07-03T09:15:23"
            ],
            "title": "Timestamp Iso"
          },
          "total_flows": {
            "description": "Total flows in frame (observed)",
            "examples": [
              190
            ],
            "title": "Total Flows",
            "type": "integer"
          }
        },
        "required": [
          "total_flows",
          "distinct_dest_ips",
          "distinct_ports",
          "predicted_malicious_count",
          "predicted_malicious_ratio"
        ],
        "title": "FlowSummary",
        "type": "object"
      },
      "Forecast": {
        "additionalProperties": true,
        "properties": {
          "alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "EWMA alpha",
            "examples": [
              0.4
            ],
            "title": "Alpha"
          },
          "engineMetadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EngineMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias camelCase"
          },
          "engine_metadata": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EngineMetadata"
              }
            ],
            "description": "Engine metadata including claim limitations"
          },
          "evidence": {
            "description": "Deterministic observed evidence for active warning - model-predicted and observable only, never raw Label",
            "examples": [
              [
                "High SYN flag rate (mean 0.42)",
                "Predicted malicious ratio 0.06 model-predicted (threshold 0.45)"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "mitre": {
            "description": "Pinned local MITRE ATT&CK mapping with cautious confidence - version 13.1",
            "examples": [
              [
                {
                  "name": "Network Service Discovery",
                  "technique_id": "T1046"
                }
              ]
            ],
            "items": {
              "$ref": "#/components/schemas/MitreTechnique"
            },
            "title": "Mitre",
            "type": "array"
          },
          "predicted_path": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Predicted dashed edge for top target - purple for model prediction",
            "examples": [
              {
                "source": "host-10",
                "status": "predicted",
                "style": "dashed",
                "target": "host-42"
              }
            ],
            "title": "Predicted Path"
          },
          "raw_risk": {
            "description": "Raw risk - learned malicious-risk probability",
            "examples": [
              0.22
            ],
            "title": "Raw Risk",
            "type": "number"
          },
          "slope": {
            "description": "Five-frame risk slope - positive trend required for warning",
            "examples": [
              0.015
            ],
            "title": "Slope",
            "type": "number"
          },
          "smoothed_risk": {
            "description": "Five-frame EWMA smoothed risk - not learned temporal world model",
            "examples": [
              0.25
            ],
            "title": "Smoothed Risk",
            "type": "number"
          },
          "stage": {
            "description": "Stage estimate transparent evidence score - not calibrated model probability",
            "examples": [
              "Normal"
            ],
            "title": "Stage",
            "type": "string"
          },
          "stage_estimate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias for stage",
            "title": "Stage Estimate"
          },
          "stage_evidence_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stage evidence score details with model-predicted ratio",
            "examples": [
              "Normal - observed evidence in frame (model-predicted ratio 0.03)"
            ],
            "title": "Stage Evidence Score"
          },
          "targetRanking": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TargetRankingEntry"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias camelCase",
            "title": "Targetranking"
          },
          "target_ranking": {
            "description": "NetworkX target-host ranking graph-ranked - not learned classifier",
            "examples": [
              [
                {
                  "host": "host-42",
                  "target_score": 12.4
                }
              ]
            ],
            "items": {
              "$ref": "#/components/schemas/TargetRankingEntry"
            },
            "title": "Target Ranking",
            "type": "array"
          },
          "threshold": {
            "description": "Frozen warning threshold chosen on Thursday validation 0.45",
            "examples": [
              0.45
            ],
            "title": "Threshold",
            "type": "number"
          },
          "warning": {
            "description": "Warning threshold crossed before ground truth requires both smoothed risk and positive slope to cross frozen thresholds",
            "examples": [
              false
            ],
            "title": "Warning",
            "type": "boolean"
          }
        },
        "required": [
          "raw_risk",
          "smoothed_risk",
          "slope",
          "warning",
          "threshold",
          "stage",
          "target_ranking",
          "evidence",
          "mitre",
          "engine_metadata"
        ],
        "title": "Forecast",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NetworkEdge": {
        "additionalProperties": true,
        "properties": {
          "activity": {
            "description": "Observed activity count in frame (observed)",
            "examples": [
              3
            ],
            "title": "Activity",
            "type": "integer"
          },
          "first_seen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Frame first seen",
            "examples": [
              0
            ],
            "title": "First Seen"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Frame last seen",
            "examples": [
              5
            ],
            "title": "Last Seen"
          },
          "novelty": {
            "description": "Edge novelty 0..1 graph-derived",
            "examples": [
              0.8
            ],
            "title": "Novelty",
            "type": "number"
          },
          "protocol": {
            "description": "Protocol TCP|UDP|ICMP",
            "examples": [
              "TCP"
            ],
            "title": "Protocol",
            "type": "string"
          },
          "protocols": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Protocols list",
            "examples": [
              [
                "TCP"
              ]
            ],
            "title": "Protocols"
          },
          "source": {
            "description": "Source host alias (observed)",
            "examples": [
              "host-1"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "description": "Status - observed for real, predicted for dashed predicted path, simulated/removed for isolation",
            "examples": [
              "observed"
            ],
            "title": "Status",
            "type": "string"
          },
          "target": {
            "description": "Target host alias (observed)",
            "examples": [
              "host-2"
            ],
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "protocol",
          "activity",
          "novelty",
          "status"
        ],
        "title": "NetworkEdge",
        "type": "object"
      },
...
      "ScenarioDetail": {
        "additionalProperties": true,
        "properties": {
          "claim_limitations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Claim limitations",
            "examples": [
              "Binary benign-versus-malicious risk only..."
            ],
            "title": "Claim Limitations"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Created at",
            "examples": [
              "2026-09-01T14:57:44.637626+00:00"
            ],
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Description",
            "examples": [
              "Approximately 30-frame replay ..."
            ],
            "title": "Description"
          },
          "engineMetadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EngineMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias camelCase"
          },
          "engine_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EngineMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Engine metadata"
          },
          "frame_count": {
            "description": "Frame count approx 30",
            "examples": [
              30
            ],
            "title": "Frame Count",
            "type": "integer"
          },
          "frames": {
            "description": "Replay frames with gated groundTruth - future labels not exposed",
            "items": {
              "$ref": "#/components/schemas/ScenarioFrame"
            },
            "title": "Frames",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Id"
          },
          "metadata": {
            "description": "Scenario metadata with disclosure, chronological, claim limitations",
            "examples": [
              {
                "disclosure": "Composite/synthetic stitching disclosed...",
                "threshold": 0.45
              }
            ],
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "description": "Human readable name",
            "examples": [
              "CyberWorld Replay - Chronological CICIDS2017 Composite"
            ],
            "title": "Name",
            "type": "string"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provenance source_url, data_mode, seed, threshold",
            "title": "Provenance"
          },
          "scenario_id": {
            "description": "Scenario ID",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Scenario Id",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deterministic seed",
            "examples": [
              42
            ],
            "title": "Seed"
          }
        },
        "required": [
          "scenario_id",
          "name",
          "frame_count",
          "frames",
          "metadata"
        ],
        "title": "ScenarioDetail",
        "type": "object"
      },
      "ScenarioFrame": {
        "additionalProperties": true,
        "properties": {
          "edges": {
            "description": "Observed edges - dashed for predicted path, muted for removed simulated path",
            "items": {
              "$ref": "#/components/schemas/NetworkEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "evidence": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Observed evidence",
            "title": "Evidence"
          },
          "flow_summary": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FlowSummary"
              }
            ],
            "description": "Flow summary observed + model-predicted fields - no raw label counts before reveal"
          },
          "forecast": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Forecast"
              }
            ],
            "description": "Forecast with stage, ranking, evidence, MITRE, engine metadata"
          },
          "frame": {
            "description": "Frame index 0..29",
            "examples": [
              5
            ],
            "title": "Frame",
            "type": "integer"
          },
          "frame_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias for frame",
            "examples": [
              5
            ],
            "title": "Frame Id"
          },
          "groundTruth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundTruth"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias camelCase"
          },
          "ground_truth": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GroundTruth"
              }
            ],
            "description": "Gated ground truth - revealed false before ground_truth_start hides label_distribution and attack_type"
          },
          "mitre": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MitreTechnique"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "MITRE mapping",
            "title": "Mitre"
          },
          "mitre_mapping": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MitreTechnique"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias MITRE",
            "title": "Mitre Mapping"
          },
          "nodes": {
            "description": "Observed host-and-edge topology nodes - cyan for observed/healthy",
            "examples": [
              [
                {
                  "criticality": "low",
                  "id": "host-1",
                  "risk": 0.2,
                  "role": "workstation",
                  "status": "observed"
                }
              ]
            ],
            "items": {
              "$ref": "#/components/schemas/NetworkNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "predicted_path": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Predicted dashed path for top target",
            "title": "Predicted Path"
          },
          "signals": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Signals"
              }
            ],
            "description": "Temporal signals raw/smoothed/slope/warning - learned raw, rule-derived smoothed/slope"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias stage estimate",
            "examples": [
              "Normal"
            ],
            "title": "Stage"
          },
          "stage_estimate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias stage",
            "examples": [
              "Normal"
            ],
            "title": "Stage Estimate"
          },
          "target_ranking": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TargetRankingEntry"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias target ranking graph-ranked",
            "title": "Target Ranking"
          },
          "timestamp": {
            "description": "Frame timestamp observed - chronological",
            "examples": [
              "2017-07-03 16:40:57"
            ],
            "title": "Timestamp",
            "type": "string"
          },
          "timestamp_iso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO timestamp",
            "examples": [
              "2017-07-03T16:40:57"
            ],
            "title": "Timestamp Iso"
          }
        },
        "required": [
          "frame",
          "timestamp",
          "flow_summary",
          "nodes",
          "edges",
          "signals",
          "forecast",
          "ground_truth"
        ],
        "title": "ScenarioFrame",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Predictive cybersecurity decision-support prototype - Replay -> Risk model -> Temporal state -> Stage estimate -> Target ranking -> Evidence and MITRE -> Simulated isolation -> Before/after comparison",
    "title": "CyberWorld AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "description": "Return scenario details with frames (timestamp, nodes, edges, signals, forecast, groundTruth gated, metadata).\n\nValidates scenario_id must exist (404 if not).\nEnsures groundTruth gating already in stored scenario (future labels not exposed before reveal).",
        "operationId": "get_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get scenario details",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  }
}